requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==3.2.16
django-phonenumber-field==7.1.0
django-smsru==1.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
packaging==23.1
phonenumbers==8.13.18
psycopg2==2.9.7
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
South==1.0.2
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
